"05-20 22:00"

+++ src/main/java/com/lzq/springbootmybatis01/service/SeckillService.java
@@ -7,5 +7,5 @@ package com.lzq.springbootmybatis01.service;
  * @create: 2020-05-20 15:28
  **/
 public interface SeckillService {
-    public String saveOrder(Long id,String userId);
+    public String saveOrder(Integer id,String userId);
 }
